Move maplike/setlike definitions to script_bindings.
